Move experimentOverriddenService code to own file
